feat: implement top rated movies retrieval in LocalMovieDataSource and update related components
