+++ app/src/main/java/com/baghdad/novix/di/RemoteDataSourceModule.kt
@@ -7,7 +7,6 @@ import com.baghdad.remoteDataSource.RemoteEpisodeDataSourceImpl
 import com.baghdad.remoteDataSource.RemoteGenreDataSourceImpl
 import com.baghdad.remoteDataSource.RemoteMovieDataSourceImpl
 import com.baghdad.remoteDataSource.RemoteSearchDataSourceImpl
-import com.baghdad.remoteDataSource.RemoteTopRatingDataSourceImpl
 import com.baghdad.remoteDataSource.RemoteTvShowDataSourceImpl
 import com.baghdad.remoteDataSource.interceptor.ApiInterceptor
 import com.baghdad.repository.datasource.remote.RemoteActorDataSource
@@ -15,7 +14,6 @@ import com.baghdad.repository.datasource.remote.RemoteEpisodeDataSource
 import com.baghdad.repository.datasource.remote.RemoteGenreDataSource
 import com.baghdad.repository.datasource.remote.RemoteMovieDataSource
 import com.baghdad.repository.datasource.remote.RemoteSearchDataSource
-import com.baghdad.repository.datasource.remote.RemoteTopRatingDataSource
 import com.baghdad.repository.datasource.remote.RemoteTvShowDataSource
 import com.baghdad.repository.language.LanguageProvider
 import io.ktor.client.HttpClient
@@ -101,15 +99,6 @@ val remoteDataSourceModule = module {
             logger = get()
         )
     }
-
-    single <RemoteTopRatingDataSource> {
-        RemoteTopRatingDataSourceImpl(
-            httpClient = get(),
-            baseUrl = get(named("BASE_URL")),
-            logger = get()
-        )
-    }
-
 
     single<LanguageProvider> { AppLanguageProvider() }
 }
