+++ local_datasource/src/main/java/com/baghdad/local_datasource/roomDB/dao/TopRatedDao.kt
@@ -11,11 +11,10 @@ interface TopRatedDao {
     suspend fun upsertMovies(movies: List<TopRatedMovie>)
 
     @Query(
-        """ SELECT * from TopRatedMovie LIMIT :pageSize OFFSET :offset"""
+        """ SELECT * from TopRatedMovie ORDER BY imdbRating DESC LIMIT :pageSize OFFSET :offset"""
     )
     suspend fun getTopRatedMoves(
         pageSize: Int,
         offset: Int,
     ): List<TopRatedMovie>
 }
-// order by after offset
