+++ repository/src/main/java/com/baghdad/repository/datasource/local/LocalMovieDataSource.kt
@@ -13,4 +13,6 @@ interface LocalMovieDataSource {
     suspend fun deleteAllMovies()
     suspend fun updateMovie(newMovie: MovieDto)
     suspend fun searchMoviesByTitle(title: String, page: Int, pageSize: Int = 20): List<MovieDto>
+    suspend fun getTopRatedMovies(page: Int, pageSize: Int): List<MovieDto>
+    suspend fun saveTopRatedMovies(movies: List<MovieDto>)
 }
