--- repository/src/main/java/com/baghdad/repository/datasource/remote/RemoteTopRatingDataSource.kt
@@ -1,8 +1,0 @@
-package com.baghdad.repository.datasource.remote
-
-import com.baghdad.repository.model.MovieDto
-import com.baghdad.repository.model.PagedResultDto
-
-interface RemoteTopRatingDataSource {
-    suspend fun getTopRatedMovies(page: Int): PagedResultDto<MovieDto>
-}
